requirements.txt
Flask==3.0.0
openai==1.3.7
requests==2.31.0
python-dotenv==1.0.0
sendgrid==6.10.0
gunicorn==21.2.0
